Thêm alertDialog phần xoá

+++ NoteApp/app/src/main/java/com/example/noteapp/MainActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
@@ -35,13 +36,15 @@ class MainActivity : AppCompatActivity() {
                 binding.closeMultiSelectButton.visibility = View.VISIBLE
                 binding.addButton.visibility = View.GONE
                 binding.deleteSelectedItemButton.visibility = View.VISIBLE
-                binding.selectedCountTextView.text = "$count selected"
-                binding.selectedCountTextView.visibility = View.VISIBLE
+//                binding.selectedCountTextView.text = "$count selected"
+//                binding.selectedCountTextView.visibility = View.VISIBLE
+                binding.notesHeading.text = "Đã chọn $count mục"
             }
             else{
                 binding.closeMultiSelectButton.visibility = View.GONE
                 binding.addButton.visibility = View.VISIBLE
-                binding.selectedCountTextView.visibility = View.GONE
+//                binding.selectedCountTextView.visibility = View.GONE
+                binding.notesHeading.text = "Notes."
                 binding.deleteSelectedItemButton.visibility = View.GONE
             }
         }
@@ -49,15 +52,24 @@ class MainActivity : AppCompatActivity() {
             notesAdapter.clearSelection()
         }
         binding.deleteSelectedItemButton.setOnClickListener {
-            val selectedPositions = notesAdapter.selectedItems.toList().sortedDescending()
+            AlertDialog.Builder(this)
+                .setTitle("Xác nhận xoá")
+                .setMessage("Bạn có chắc chắn muốn xoá các mục đã chọn?")
+                .setPositiveButton("Xoá") { _, _ ->
+                    val selectedPositions = notesAdapter.selectedItems.toList().sortedDescending()
 
-            for (index in selectedPositions) {
-                val noteId = notesAdapter.notes[index].id
-                db.deleteNote(noteId)
-            }
+                    for (index in selectedPositions) {
+                        val noteId = notesAdapter.notes[index].id
+                        db.deleteNote(noteId)
+                    }
 
-            notesAdapter.refreshData(db.getAllNotes())
-            notesAdapter.clearSelection()
+                    notesAdapter.refreshData(db.getAllNotes())
+                    notesAdapter.clearSelection()
+                }
+                .setNegativeButton("Huỷ") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .show()
         }
     }
